requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
jinja2>=3.1.0
pytest>=7.4.0
flask>=2.3.0
flask-cors>=4.0.0
